requirements.txt
streamlit
azure-search-documents
azure-identity
langchain
langchain-openai
langchain-community
python-dotenv
